% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\docType{data}
\name{b_optimize}
\alias{b_optimize}
\title{Compute the volume of champagne poured in a glass using "optimize".

Define the volume of champagne poured in a glass using "optimize".}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
b_optimize
}
\arguments{
\item{x}{A list or atomic vector.}

\item{f}{A function.}
}
\value{
The volume of champagne poured in a glass.
}
\description{
Compute the volume of champagne poured in a glass using "optimize".

Define the volume of champagne poured in a glass using "optimize".
}
\examples{
#' # Define the "integrand" function
f <- function(x) {
result <- numeric(length(x))
for (i in seq_along(x)) {
  result[i] <- f(x[i])
}
return(result)
}

integrand <- function(x) {
sapply(x, f)
}
}
\author{
Adlan Boithier
}
\keyword{datasets}
