% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extracting results of BM.R
\docType{data}
\name{benchmark_data}
\alias{benchmark_data}
\title{Extracting the results of a benchmark.

Extract the results of a benchmark.}
\format{
An object of class \code{data.frame} with 1600 rows and 3 columns.
}
\usage{
benchmark_data
}
\arguments{
\item{benchmarking_results}{A list containing the results of benchmarking different functions.}
}
\value{
The result of a benchmark test.
}
\description{
Extracting the results of a benchmark.

Extract the results of a benchmark.
}
\examples{
# Assuming you have benchmarked some functions and obtained the results
# using the benchmarking_results function.
# Example benchmarking_results function call:
# input <- c(10, 100, 1000, 10000)
# functions <- list(f_loop, f_purrr, f_sapply, f_Vectorize)
# benchmarking_results(input, functions)

# Now, you can extract the results using benchmark_data function:
benchmark_data <- extract_benchmark_results(benchmarking_results)
}
\author{
Adlan Boithier
}
\keyword{datasets}
