% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{champagne_b_optimize}
\alias{champagne_b_optimize}
\title{Compute volume using optimize.

Define the volume of champagne poured in a glass using "optimize".}
\usage{
champagne_b_optimize(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of champagne poured in a glass.
}
\description{
Compute volume using optimize.

Define the volume of champagne poured in a glass using "optimize".
}
\author{
Adlan Boithier
}
