% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot.R
\docType{data}
\name{champagne_b_uniroot}
\alias{champagne_b_uniroot}
\title{Compute volume using uniroot.

Define the volume of champagne poured in a glass using "uniroot".}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
champagne_b_uniroot(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of champagne poured in a glass.
}
\description{
Compute volume using uniroot.

Define the volume of champagne poured in a glass using "uniroot".
}
\examples{
#' # Define the "integrand" function
f <- function(x) {
result <- numeric(length(x))
for (i in seq_along(x)) {
  result[i] <- champagne_f(x[i])
}
return(result)
}

integrand <- function(x) {
sapply(x, champagne_f)
}
}
\author{
Adlan Boithier
}
\keyword{datasets}
