% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non-vectorized_f.R
\name{champagne_f}
\alias{champagne_f}
\title{Non-vectorized f

Non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
champagne_f(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Non-vectorized f

Non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- champagne_f(2)
}
\author{
Adlan Boithier
}
