% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For_loop_f.R
\name{champagne_f_loop}
\alias{champagne_f_loop}
\title{Vectorize f with a for loop

Vectorize, by using a "for" loop, the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
champagne_f_loop(x)
}
\arguments{
\item{x}{A list or atomic vector}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorize f with a for loop

Vectorize, by using a "for" loop, the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- champagne_f_loop(2)
}
\author{
Adlan Boithier
}
