% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_f.R
\name{champagne_f_purrr}
\alias{champagne_f_purrr}
\title{Vectorize f with purrr

Vectorize, by using a "map" function of "purrr", the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
champagne_f_purrr(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{champagne_f}{A function.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorize f with purrr

Vectorize, by using a "map" function of "purrr", the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- champagne_f_purrr(c(1,2,3))
}
\author{
Adlan Boithier
}
