% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_f.R
\name{champagne_f_sapply}
\alias{champagne_f_sapply}
\title{Vectorization of f with apply.

Vectorize, by using one of the "apply" functions.", the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
champagne_f_sapply(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{champagne_f}{A function.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorization of f with apply.

Vectorize, by using one of the "apply" functions.", the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- champagne_f_sapply(c(1,2,3))
}
\author{
Adlan Boithier
}
