% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectorize_f.R
\name{champagne_f_Vectorize}
\alias{champagne_f_Vectorize}
\title{Vectorize f by using Vectorize

Vectorize, by using "Vectorize", the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
champagne_f_Vectorize(x)
}
\arguments{
\item{champagne_f}{A function.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorize f by using Vectorize

Vectorize, by using "Vectorize", the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- champagne_f_Vectorize(f)
}
\author{
Adlan Boithier
}
