% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectorized_f.R
\name{champagne_f_vectorized}
\alias{champagne_f_vectorized}
\title{Vectorized f

Vectorized function to compute the volume of a cup of Champagne glass}
\usage{
champagne_f_vectorized(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorized f

Vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- champagne_f_vectorized(2)
}
\author{
Adlan Boithier
}
