% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrand.R
\name{champagne_integrand}
\alias{champagne_integrand}
\title{Integrand for volume calculation

Define the integrand for volume calculation.}
\usage{
champagne_integrand(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{champagne_f}{A function.}
}
\value{
The integrand.
}
\description{
Integrand for volume calculation

Define the integrand for volume calculation.
}
\examples{
# Define the "f_loop" function
champagne_f_loop <- function(x) {
result <- numeric(length(x))
for (i in seq_along(x)) {
  result[i] <- f(x[i])
}
return(result)
}

}
\author{
Adlan Boithier
}
