% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parties.R
\name{champagne_Parties}
\alias{champagne_Parties}
\title{Simulate parties and estimate Champagne needed.

This function simulates parties and estimates the amount of Champagne needed based on various parameters.}
\usage{
champagne_Parties(n, lambda_G, lambda_D, a, mu_b, sd_b, volume_bottle)
}
\arguments{
\item{n}{The number of parties to simulate.}

\item{lambda_G}{Lambda of the Poisson Distribution regarding the number of guests.}

\item{lambda_D}{Lambda of the Poisson Distribution regarding the number of glasses drank by a guest.}

\item{a}{Bottom of the glass.}

\item{mu_b}{Mean of the normal distribution.}

\item{sd_b}{Standard deviation of the normal distribution.}

\item{volume_bottle}{The volume of a Champagne bottle.}
}
\value{
A tibble with columns 'party' (party number), 'G' (number of guests), 'D' (total glasses drank),
'b' (total Champagne in milliliters), and 'bottles' (number of Champagne bottles needed).
}
\description{
Simulate parties and estimate Champagne needed.

This function simulates parties and estimates the amount of Champagne needed based on various parameters.
}
\examples{
\dontrun{
# Example function call:
result <- champagne_Parties(n = 100, lambda_G = 10, lambda_D = 5, a = 0.5, mu_b = 5, sd_b = 2, volume_bottle = 0.75)
}

}
\author{
Adlan Boithier
}
