% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function f non-vectorized.R
\name{f}
\alias{f}
\title{Non-vectorized version of f

Non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
f(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Non-vectorized version of f

Non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- f(2)
}
\author{
Adlan Boithier
}
