% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function f "for" loop.R
\name{f_loop}
\alias{f_loop}
\title{Vectorization of f by using a "for" loop

Vectorize, by using a "for" loop, the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
f_loop(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorization of f by using a "for" loop

Vectorize, by using a "for" loop, the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- f_loop(2)
}
\author{
Adlan Boithier
}
