% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function f purrr.R
\name{f_purrr}
\alias{f_purrr}
\title{Vectorization of f by using a "map" function of "purrr"

Vectorize, by using a "map" function of "purrr", the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
f_purrr(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{f}{A function.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorization of f by using a "map" function of "purrr"

Vectorize, by using a "map" function of "purrr", the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- f_purrr(2, f)
}
\author{
Adlan Boithier
}
