% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function f apply.R
\name{f_sapply}
\alias{f_sapply}
\title{Vectorization of f by using one of the "apply" functions.

Vectorize, by using one of the "apply" functions.", the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
f_sapply(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{f}{A function.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorization of f by using one of the "apply" functions.

Vectorize, by using one of the "apply" functions.", the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- f_sapply(2, f)
}
\author{
Adlan Boithier
}
