% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function f Vectorize.R
\name{f_Vectorize}
\alias{f_Vectorize}
\title{Vectorization of f by using a "Vectorize"

Vectorize, by using "Vectorize", the non-vectorized function to compute the volume of a cup of Champagne glass}
\usage{
f_Vectorize(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{f}{A function.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorization of f by using a "Vectorize"

Vectorize, by using "Vectorize", the non-vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- f_Vectorize(f)
}
\author{
Adlan Boithier
}
