% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectorized version of f.R
\name{f_vectorized}
\alias{f_vectorized}
\title{Vectorized version of f

Vectorized function to compute the volume of a cup of Champagne glass}
\usage{
f_vectorized(x)
}
\arguments{
\item{x}{A list or atomic vector.}
}
\value{
The volume of a cup of Champagne glass
}
\description{
Vectorized version of f

Vectorized function to compute the volume of a cup of Champagne glass
}
\examples{
f1 <- f_vectorized(2)
}
\author{
Adlan Boithier
}
