% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrand.R
\name{integrand}
\alias{integrand}
\title{Compute the integrand for volume calculation.

Define the integrand for volume calculation.}
\usage{
integrand(x)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{f}{A function.}
}
\value{
The integrand.
}
\description{
Compute the integrand for volume calculation.

Define the integrand for volume calculation.
}
\examples{
#' # Define the "f_loop" function
f <- function(x) {
result <- numeric(length(x))
for (i in seq_along(x)) {
  result[i] <- f(x[i])
}
return(result)
}

}
\author{
Adlan Boithier
}
