% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\docType{data}
\name{champagne_b_optimize}
\alias{champagne_b_optimize}
\title{Compute volume using optimize.

Define the volume of champagne poured in a glass using "optimize".}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
champagne_b_optimize
}
\arguments{
\item{x}{A list or atomic vector.}

\item{champagne_f}{A function.}
}
\value{
The volume of champagne poured in a glass.
}
\description{
Compute volume using optimize.

Define the volume of champagne poured in a glass using "optimize".
}
\examples{
#' # Define the "integrand" function
champagne_integrand <- function(x) {
result <- numeric(length(x))
for (i in seq_along(x)) {
  result[i] <- champagne_f(x[i])
}
return(result)
}

champagne_integrand <- function(x) {
sapply(x, champagne_f)
}
}
\author{
Adlan Boithier
}
\keyword{datasets}
